Correctly discard cards on the server side
The original discard behavior only called the `removeCardFromHand` function to discard a card, but this function only removes the DOM elements associated with the discarded card, but not the underlying player hand state on the server side. This led to fewer than 10 cards being redrawn at the start of the next round, since these "invisible" cards still took up space in the player's hand.

To get the server to correctly add the discarded card to the "discard pile" and fully remove it from the player's hand, a new AJAX operation `DISCARD_CARD` was made (based on the similar `PLAY_CARD` operation).

**Discard button successful code path:**
1. Pressing the `Discard` button flips on the `discardMode_` flag via the `toggleDiscardMode_` event handler in `cah.game.js`.
2. A card in the hand is clicked, which triggers a `cah.$.AjaxOperation.DISCARD_CARD` call via the `handCardClick_` handler in `cah.game.js`. This signals to the server that a card should be discarded.
3. The AJAX call is caught by `DiscardCardHandler.java`, which then calls the `discardCard` function within `Game.java`.
4. The `discardCard` function calls `whiteDeck.discard()`, which moves the card to the server's discard pile.
5. The `DiscardCardHandler` then passes an `ErrorCode` (`null` when successful) to the handler defined in `cah.ajax.handlers.js` as `cah.ajax.SuccessHandlers[cah.$.AjaxOperation.DISCARD_CARD]`.
6. The success handler then calls `discardCardComplete` in `cah.game.js` to handle the client side effects of the discard.