Use a class variable to keep track of the discard card
The original `this.removeCardFromHand(card);` line was copied from the `handCardClick_` function but the reference to `card` is not defined in the new scope.
A class variable `markedForDiscard_` is created to keep track of the client-side card elements across the AJAX call, so they can be correctly removed in the `discardCardComplete` function call at the end.